refactor: split bounty-detail-submissions-card into smaller components
Extract row, dialog, and helper modules under components/bounty-detail/submissions/ while preserving the existing import path via a barrel re-export.
